+++ api/sources.go
@@ -22,8 +22,14 @@ func InitSourcesRouter(apiHelper Helper) {
 		sourcename := r.Param("sourcename")
 		json, count, err := database.GetRecords(c, sourcename, r.URL.Query())
 		if err == nil {
-			SetResponseHeaders(c, w, r.Request, count)
-			return WriteContent(c, w, http.StatusOK, json)
+			status := SetResponseHeaders(c, w, r, count)
+			if status >= http.StatusBadRequest {
+				return heligo.WriteHeader(w, status)
+			}
+			if status == 0 {
+				status = http.StatusOK
+			}
+			return WriteContent(c, w, status, json)
 		} else {
 			return WriteError(w, err)
 		}
@@ -37,6 +43,7 @@ func InitSourcesRouter(apiHelper Helper) {
 		}
 		data, count, err := database.CreateRecords(c, sourcename, records, r.URL.Query())
 		if err == nil {
+			SetResponseHeaders(c, w, r, count)
 			if data == nil {
 				return heligo.WriteJSON(w, http.StatusCreated, count)
 			} else {
@@ -55,10 +62,10 @@ func InitSourcesRouter(apiHelper Helper) {
 		}
 		data, count, err := database.UpdateRecords(c, sourcename, records[0], r.URL.Query())
 		if err == nil {
+			SetResponseHeaders(c, w, r, count)
 			if data == nil {
 				return heligo.WriteHeader(w, http.StatusNoContent)
 			} else {
-				SetResponseHeaders(c, w, r.Request, count)
 				return WriteContent(c, w, http.StatusOK, data)
 			}
 		} else {
@@ -68,8 +75,9 @@ func InitSourcesRouter(apiHelper Helper) {
 
 	api.Handle("DELETE", "/:sourcename", func(c context.Context, w http.ResponseWriter, r heligo.Request) (int, error) {
 		sourcename := r.Param("sourcename")
-		data, _, err := database.DeleteRecords(c, sourcename, r.URL.Query())
+		data, count, err := database.DeleteRecords(c, sourcename, r.URL.Query())
 		if err == nil {
+			SetResponseHeaders(c, w, r, count)
 			if data == nil {
 				return heligo.WriteHeader(w, http.StatusNoContent)
 			} else {
@@ -86,8 +94,14 @@ func InitSourcesRouter(apiHelper Helper) {
 		fname := r.Param("fname")
 		json, count, err := database.ExecFunction(c, fname, nil, r.URL.Query(), true)
 		if err == nil {
-			SetResponseHeaders(c, w, r.Request, count)
-			return WriteContent(c, w, http.StatusOK, json)
+			status := SetResponseHeaders(c, w, r, count)
+			if status >= http.StatusBadRequest {
+				return heligo.WriteHeader(w, status)
+			}
+			if status == 0 {
+				status = http.StatusOK
+			}
+			return WriteContent(c, w, status, json)
 		} else {
 			return WriteError(w, err)
 		}
@@ -101,10 +115,10 @@ func InitSourcesRouter(apiHelper Helper) {
 		}
 		data, count, err := database.ExecFunction(c, fname, records[0], r.URL.Query(), false)
 		if err == nil {
+			SetResponseHeaders(c, w, r, count)
 			if data == nil {
 				return heligo.WriteJSON(w, http.StatusOK, count)
 			} else {
-				SetResponseHeaders(c, w, r.Request, count)
 				return WriteContent(c, w, http.StatusOK, data)
 			}
 		} else {